removed preferences selection builder from drawer to separate component
